Lots of minor fixes. Details below.
CRUD actions for events and expenses are now connected to controllers, $event argument is used to parse id of the element to controller.
Added ng-controller directoves to all the divs, fixed invalid references.
Implemented a reference to selectedItems.user from expenses and events controllers for testing purposes.
Added a function that clears selected car's active status when 'all cars' button is clicked.